image: Reset Private Image Data After Executing Test
Currently, the image.rs tests don't reset the private image
data after executing tests. This means that on the next test
execution, the GCD gets reset first, then the private image data
reset, which causes the memory it holds to get dropped, which then
will fail to be freed against the now empty GCD.

With reworks to the image code to add debug_asserts to ensure the
memory is being freed correctly, this patterns faults.

This changes the test to clean up the private image data while the
GCD is still valid.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>