Memory Attribute Protocol: Fix get_memory_attributes
get_memory_attributes currently enforces that the range
being queried exactly matches a GCD descriptor. This is
incorrect, it can be a subset of a descriptor or span
descriptors, if the attributes are the same.

This fixes the logic to use the for each descriptor function
and ensure the attributes are the same across the range,
returning an error if not.

This also adds unit tests to the file, bringing code cov
from 0% -> 92%.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>